+++ internal/service/taskservice/service.go
@@ -7,9 +7,9 @@ import (
 
 type Repository interface {
 	Create(u entity.Task) (entity.Task, error)
-	GetByID(id uint) (entity.Task, error)
-	GetAllByUserID(userID uint) ([]entity.Task, error)
-	IsExistsUser(id uint) (bool, error)
+	GetByID(id uint64) (entity.Task, error)
+	GetAllByUserID(userID uint64) ([]entity.Task, error)
+	IsExistsUser(id uint64) (bool, error)
 	// etc
 }
